scrapers/weather_scraper.py
"""Weather data scraper for IMD (India Meteorological Department)."""
from .base_scraper import BaseScraper
from bs4 import BeautifulSoup
from typing import Dict, Any, List
import pandas as pd
from datetime import datetime, timedelta

class IMDWeatherScraper(BaseScraper):
    def __init__(self):
        super().__init__(base_url="https://mausam.imd.gov.in")
        self.headers = {
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
        }

    def get_weather_data(self, region: str = "Gujarat") -> Dict[str, Any]:
        """
        Fetch weather data for a specific region.
        Focus on key agricultural weather parameters.
        """
        try:
            # Example endpoint - actual endpoints will need to be identified
            response = self.make_request(
                endpoint="/weather_forecast",
                params={"region": region},
                headers=self.headers
            )
            
            soup = BeautifulSoup(response.text, 'lxml')
            
            # Extract weather data (example structure - adjust based on actual website)
            weather_data = {
                "timestamp": datetime.now().isoformat(),
                "region": region,
                "temperature": self._extract_temperature(soup),
                "rainfall": self._extract_rainfall(soup),
                "humidity": self._extract_humidity(soup),
                "forecast": self._extract_forecast(soup)
            }
            
            return self.save_data(weather_data, "weather")
            
        except Exception as e:
            self.logger.error(f"Error fetching weather data for {region}: {str(e)}")
            raise

    def _extract_temperature(self, soup: BeautifulSoup) -> Dict[str, float]:
        """Extract temperature data from the page."""
        # Implementation will depend on actual HTML structure
        return {
            "current": 0.0,  # Replace with actual scraping logic
            "min": 0.0,
            "max": 0.0
        }

    def _extract_rainfall(self, soup: BeautifulSoup) -> Dict[str, float]:
        """Extract rainfall data from the page."""
        # Implementation will depend on actual HTML structure
        return {
            "daily": 0.0,  # Replace with actual scraping logic
            "weekly_forecast": []
        }

    def _extract_humidity(self, soup: BeautifulSoup) -> float:
        """Extract humidity data from the page."""
        # Implementation will depend on actual HTML structure
        return 0.0  # Replace with actual scraping logic

    def _extract_forecast(self, soup: BeautifulSoup) -> List[Dict[str, Any]]:
        """Extract weather forecast for next few days."""
        # Implementation will depend on actual HTML structure
        forecast = []
        # Add logic to extract 5-day forecast
        return forecast 