requirements.txt
Flask==2.3.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
flask-restx==1.1.0
SQLAlchemy==2.0.23
alembic==1.13.0
pandas>=2.1.3
numpy>=1.26.0
scikit-learn>=1.3.2
xgboost>=2.0.2
requests==2.31.0
python-dotenv==1.0.0
gunicorn==20.1.0
beautifulsoup4==4.12.2
lxml==4.9.3
joblib>=1.2.0
aiohttp==3.9.1
schedule==1.2.1
selenium==4.15.2
webdriver-manager==4.0.1
python-dateutil==2.8.2
APScheduler==3.10.4
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0